v0.4.12 beta: Fixed the saving!
I finally found the bug that has plagued this game for over a year! Basically, whenever you played Elemental Incremental, when Elemental Incremental automatically saved, it would overwrite this game's save data! Then upon loading this game's data, it wouldn't work because it was fed the save file of Elemental Incremental, which is garbage in this game. Wow, I thought for all this time it was some browser update deprecating a feature used in the code of this game.